chore: decouple parser utility
